requirements.txt
# For dbt 0.19.0, it requires these versions
google-cloud-bigquery<2.4,>=1.25.0
google-cloud-core<1.5,>=1.3.0
google-api-core<1.24,>=1.16.0

dbt==0.19.0